deps/v8/test/webkit/fast/js/toString-number-expected.txt

# Copyright 2013 the V8 project authors. All rights reserved.
# Copyright (C) 2005, 2006, 2007, 2008, 2009 Apple Inc. All rights reserved.
#
# Redistribution and use in source and binary forms, with or without
# modification, are permitted provided that the following conditions
# are met:
# 1.  Redistributions of source code must retain the above copyright
#     notice, this list of conditions and the following disclaimer.
# 2.  Redistributions in binary form must reproduce the above copyright
#     notice, this list of conditions and the following disclaimer in the
#     documentation and/or other materials provided with the distribution.
#
# THIS SOFTWARE IS PROVIDED BY APPLE INC. AND ITS CONTRIBUTORS ``AS IS'' AND ANY
# EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
# WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
# DISCLAIMED. IN NO EVENT SHALL APPLE INC. OR ITS CONTRIBUTORS BE LIABLE FOR ANY
# DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
# (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
# LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
# ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
# (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
# SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Test the conversion performed by the function Number.prototype.toString.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Number(0).toString() is "0"
PASS Number.prototype.toString.call(0) is "0"
PASS Number.prototype.toString.call(new Number(0)) is "0"
PASS Number("0").toString() is "0"
PASS Number(0).toString(10) is "0"
PASS Number(0).toString(2) is "0"
PASS Number.prototype.toString.call(0, 2) is "0"
PASS Number.prototype.toString.call(new Number(0), 2) is "0"
PASS Number(0).toString(36) is "0"
PASS Number.prototype.toString.call(0, 36) is "0"
PASS Number.prototype.toString.call(new Number(0), 36) is "0"
PASS Number(-1).toString() is "-1"
PASS Number.prototype.toString.call(-1) is "-1"
PASS Number.prototype.toString.call(new Number(-1)) is "-1"
PASS Number("-1").toString() is "-1"
PASS Number(-1).toString(10) is "-1"
PASS Number(-1).toString(2) is "-1"
PASS Number.prototype.toString.call(-1, 2) is "-1"
PASS Number.prototype.toString.call(new Number(-1), 2) is "-1"
PASS Number(-1).toString(36) is "-1"
PASS Number.prototype.toString.call(-1, 36) is "-1"
PASS Number.prototype.toString.call(new Number(-1), 36) is "-1"
PASS Number(1).toString() is "1"
PASS Number.prototype.toString.call(1) is "1"
PASS Number.prototype.toString.call(new Number(1)) is "1"
PASS Number("1").toString() is "1"
PASS Number(1).toString(10) is "1"
PASS Number(1).toString(2) is "1"
PASS Number.prototype.toString.call(1, 2) is "1"
PASS Number.prototype.toString.call(new Number(1), 2) is "1"
PASS Number(1).toString(36) is "1"
PASS Number.prototype.toString.call(1, 36) is "1"
PASS Number.prototype.toString.call(new Number(1), 36) is "1"
PASS Number(1984).toString() is "1984"
PASS Number.prototype.toString.call(1984) is "1984"
PASS Number.prototype.toString.call(new Number(1984)) is "1984"
PASS Number("1984").toString() is "1984"
PASS Number(1984).toString(10) is "1984"
PASS Number(1984).toString(2) is "11111000000"
PASS Number.prototype.toString.call(1984, 2) is "11111000000"
PASS Number.prototype.toString.call(new Number(1984), 2) is "11111000000"
PASS Number(1984).toString(36) is "1j4"
PASS Number.prototype.toString.call(1984, 36) is "1j4"
PASS Number.prototype.toString.call(new Number(1984), 36) is "1j4"
PASS Number(-1984).toString() is "-1984"
PASS Number.prototype.toString.call(-1984) is "-1984"
PASS Number.prototype.toString.call(new Number(-1984)) is "-1984"
PASS Number("-1984").toString() is "-1984"
PASS Number(-1984).toString(10) is "-1984"
PASS Number(-1984).toString(2) is "-11111000000"
PASS Number.prototype.toString.call(-1984, 2) is "-11111000000"
PASS Number.prototype.toString.call(new Number(-1984), 2) is "-11111000000"
PASS Number(-1984).toString(36) is "-1j4"
PASS Number.prototype.toString.call(-1984, 36) is "-1j4"
PASS Number.prototype.toString.call(new Number(-1984), 36) is "-1j4"
PASS Number(2147483647).toString() is "2147483647"
PASS Number.prototype.toString.call(2147483647) is "2147483647"
PASS Number.prototype.toString.call(new Number(2147483647)) is "2147483647"
PASS Number("2147483647").toString() is "2147483647"
PASS Number(2147483647).toString(10) is "2147483647"
PASS Number(2147483647).toString(2) is "1111111111111111111111111111111"
PASS Number.prototype.toString.call(2147483647, 2) is "1111111111111111111111111111111"
PASS Number.prototype.toString.call(new Number(2147483647), 2) is "1111111111111111111111111111111"
PASS Number(2147483647).toString(36) is "zik0zj"
PASS Number.prototype.toString.call(2147483647, 36) is "zik0zj"
PASS Number.prototype.toString.call(new Number(2147483647), 36) is "zik0zj"
PASS Number(-2147483648).toString() is "-2147483648"
PASS Number.prototype.toString.call(-2147483648) is "-2147483648"
PASS Number.prototype.toString.call(new Number(-2147483648)) is "-2147483648"
PASS Number("-2147483648").toString() is "-2147483648"
PASS Number(-2147483648).toString(10) is "-2147483648"
PASS Number(-2147483648).toString(2) is "-10000000000000000000000000000000"
PASS Number.prototype.toString.call(-2147483648, 2) is "-10000000000000000000000000000000"
PASS Number.prototype.toString.call(new Number(-2147483648), 2) is "-10000000000000000000000000000000"
PASS Number(-2147483648).toString(36) is "-zik0zk"
PASS Number.prototype.toString.call(-2147483648, 36) is "-zik0zk"
PASS Number.prototype.toString.call(new Number(-2147483648), 36) is "-zik0zk"
PASS Number(9007199254740992).toString() is "9007199254740992"
PASS Number.prototype.toString.call(9007199254740992) is "9007199254740992"
PASS Number.prototype.toString.call(new Number(9007199254740992)) is "9007199254740992"
PASS Number("9007199254740992").toString() is "9007199254740992"
PASS Number(9007199254740992).toString(10) is "9007199254740992"
PASS Number(9007199254740992).toString(2) is "100000000000000000000000000000000000000000000000000000"
PASS Number.prototype.toString.call(9007199254740992, 2) is "100000000000000000000000000000000000000000000000000000"
PASS Number.prototype.toString.call(new Number(9007199254740992), 2) is "100000000000000000000000000000000000000000000000000000"
PASS Number(9007199254740992).toString(36) is "2gosa7pa2gw"
PASS Number.prototype.toString.call(9007199254740992, 36) is "2gosa7pa2gw"
PASS Number.prototype.toString.call(new Number(9007199254740992), 36) is "2gosa7pa2gw"
PASS Number(-9007199254740992).toString() is "-9007199254740992"
PASS Number.prototype.toString.call(-9007199254740992) is "-9007199254740992"
PASS Number.prototype.toString.call(new Number(-9007199254740992)) is "-9007199254740992"
PASS Number("-9007199254740992").toString() is "-9007199254740992"
PASS Number(-9007199254740992).toString(10) is "-9007199254740992"
PASS Number(-9007199254740992).toString(2) is "-100000000000000000000000000000000000000000000000000000"
PASS Number.prototype.toString.call(-9007199254740992, 2) is "-100000000000000000000000000000000000000000000000000000"
PASS Number.prototype.toString.call(new Number(-9007199254740992), 2) is "-100000000000000000000000000000000000000000000000000000"
PASS Number(-9007199254740992).toString(36) is "-2gosa7pa2gw"
PASS Number.prototype.toString.call(-9007199254740992, 36) is "-2gosa7pa2gw"
PASS Number.prototype.toString.call(new Number(-9007199254740992), 36) is "-2gosa7pa2gw"
PASS Number(0).toString() is "0"
PASS Number.prototype.toString.call(0) is "0"
PASS Number.prototype.toString.call(new Number(0)) is "0"
PASS Number("0").toString() is "0"
PASS Number(0).toString(10) is "0"
PASS Number(0).toString(2) is "0"
PASS Number.prototype.toString.call(0, 2) is "0"
PASS Number.prototype.toString.call(new Number(0), 2) is "0"
PASS Number(0).toString(36) is "0"
PASS Number.prototype.toString.call(0, 36) is "0"
PASS Number.prototype.toString.call(new Number(0), 36) is "0"
PASS Number(-1).toString() is "-1"
PASS Number.prototype.toString.call(-1) is "-1"
PASS Number.prototype.toString.call(new Number(-1)) is "-1"
PASS Number("-1").toString() is "-1"
PASS Number(-1).toString(10) is "-1"
PASS Number(-1).toString(2) is "-1"
PASS Number.prototype.toString.call(-1, 2) is "-1"
PASS Number.prototype.toString.call(new Number(-1), 2) is "-1"
PASS Number(-1).toString(36) is "-1"
PASS Number.prototype.toString.call(-1, 36) is "-1"
PASS Number.prototype.toString.call(new Number(-1), 36) is "-1"
PASS Number(1).toString() is "1"
PASS Number.prototype.toString.call(1) is "1"
PASS Number.prototype.toString.call(new Number(1)) is "1"
PASS Number("1").toString() is "1"
PASS Number(1).toString(10) is "1"
PASS Number(1).toString(2) is "1"
PASS Number.prototype.toString.call(1, 2) is "1"
PASS Number.prototype.toString.call(new Number(1), 2) is "1"
PASS Number(1).toString(36) is "1"
PASS Number.prototype.toString.call(1, 36) is "1"
PASS Number.prototype.toString.call(new Number(1), 36) is "1"
PASS Number(1984).toString() is "1984"
PASS Number.prototype.toString.call(1984) is "1984"
PASS Number.prototype.toString.call(new Number(1984)) is "1984"
PASS Number("1984").toString() is "1984"
PASS Number(1984).toString(10) is "1984"
PASS Number(1984).toString(2) is "11111000000"
PASS Number.prototype.toString.call(1984, 2) is "11111000000"
PASS Number.prototype.toString.call(new Number(1984), 2) is "11111000000"
PASS Number(1984).toString(36) is "1j4"
PASS Number.prototype.toString.call(1984, 36) is "1j4"
PASS Number.prototype.toString.call(new Number(1984), 36) is "1j4"
PASS Number(-1984).toString() is "-1984"
PASS Number.prototype.toString.call(-1984) is "-1984"
PASS Number.prototype.toString.call(new Number(-1984)) is "-1984"
PASS Number("-1984").toString() is "-1984"
PASS Number(-1984).toString(10) is "-1984"
PASS Number(-1984).toString(2) is "-11111000000"
PASS Number.prototype.toString.call(-1984, 2) is "-11111000000"
PASS Number.prototype.toString.call(new Number(-1984), 2) is "-11111000000"
PASS Number(-1984).toString(36) is "-1j4"
PASS Number.prototype.toString.call(-1984, 36) is "-1j4"
PASS Number.prototype.toString.call(new Number(-1984), 36) is "-1j4"
PASS Number(2147483647).toString() is "2147483647"
PASS Number.prototype.toString.call(2147483647) is "2147483647"
PASS Number.prototype.toString.call(new Number(2147483647)) is "2147483647"
PASS Number("2147483647").toString() is "2147483647"
PASS Number(2147483647).toString(10) is "2147483647"
PASS Number(2147483647).toString(2) is "1111111111111111111111111111111"
PASS Number.prototype.toString.call(2147483647, 2) is "1111111111111111111111111111111"
PASS Number.prototype.toString.call(new Number(2147483647), 2) is "1111111111111111111111111111111"
PASS Number(2147483647).toString(36) is "zik0zj"
PASS Number.prototype.toString.call(2147483647, 36) is "zik0zj"
PASS Number.prototype.toString.call(new Number(2147483647), 36) is "zik0zj"
PASS Number(-2147483648).toString() is "-2147483648"
PASS Number.prototype.toString.call(-2147483648) is "-2147483648"
PASS Number.prototype.toString.call(new Number(-2147483648)) is "-2147483648"
PASS Number("-2147483648").toString() is "-2147483648"
PASS Number(-2147483648).toString(10) is "-2147483648"
PASS Number(-2147483648).toString(2) is "-10000000000000000000000000000000"
PASS Number.prototype.toString.call(-2147483648, 2) is "-10000000000000000000000000000000"
PASS Number.prototype.toString.call(new Number(-2147483648), 2) is "-10000000000000000000000000000000"
PASS Number(-2147483648).toString(36) is "-zik0zk"
PASS Number.prototype.toString.call(-2147483648, 36) is "-zik0zk"
PASS Number.prototype.toString.call(new Number(-2147483648), 36) is "-zik0zk"
PASS Number(9007199254740992).toString() is "9007199254740992"
PASS Number.prototype.toString.call(9007199254740992) is "9007199254740992"
PASS Number.prototype.toString.call(new Number(9007199254740992)) is "9007199254740992"
PASS Number("9007199254740992").toString() is "9007199254740992"
PASS Number(9007199254740992).toString(10) is "9007199254740992"
PASS Number(9007199254740992).toString(2) is "100000000000000000000000000000000000000000000000000000"
PASS Number.prototype.toString.call(9007199254740992, 2) is "100000000000000000000000000000000000000000000000000000"
PASS Number.prototype.toString.call(new Number(9007199254740992), 2) is "100000000000000000000000000000000000000000000000000000"
PASS Number(9007199254740992).toString(36) is "2gosa7pa2gw"
PASS Number.prototype.toString.call(9007199254740992, 36) is "2gosa7pa2gw"
PASS Number.prototype.toString.call(new Number(9007199254740992), 36) is "2gosa7pa2gw"
PASS Number(-9007199254740992).toString() is "-9007199254740992"
PASS Number.prototype.toString.call(-9007199254740992) is "-9007199254740992"
PASS Number.prototype.toString.call(new Number(-9007199254740992)) is "-9007199254740992"
PASS Number("-9007199254740992").toString() is "-9007199254740992"
PASS Number(-9007199254740992).toString(10) is "-9007199254740992"
PASS Number(-9007199254740992).toString(2) is "-100000000000000000000000000000000000000000000000000000"
PASS Number.prototype.toString.call(-9007199254740992, 2) is "-100000000000000000000000000000000000000000000000000000"
PASS Number.prototype.toString.call(new Number(-9007199254740992), 2) is "-100000000000000000000000000000000000000000000000000000"
PASS Number(-9007199254740992).toString(36) is "-2gosa7pa2gw"
PASS Number.prototype.toString.call(-9007199254740992, 36) is "-2gosa7pa2gw"
PASS Number.prototype.toString.call(new Number(-9007199254740992), 36) is "-2gosa7pa2gw"
PASS Number(0.1).toString() is "0.1"
PASS Number.prototype.toString.call(0.1) is "0.1"
PASS Number.prototype.toString.call(new Number(0.1)) is "0.1"
PASS Number("0.1").toString() is "0.1"
PASS Number(0.1).toString(10) is "0.1"
PASS Number(0.1).toString(2) is "0.0001100110011001100110011001100110011001100110011001101"
PASS Number.prototype.toString.call(0.1, 2) is "0.0001100110011001100110011001100110011001100110011001101"
PASS Number.prototype.toString.call(new Number(0.1), 2) is "0.0001100110011001100110011001100110011001100110011001101"
FAIL Number(0.1).toString(36) should be 0.3lllllllllm. Was 0.3llllllllllqsn8td1p464unmi.
FAIL Number.prototype.toString.call(0.1, 36) should be 0.3lllllllllm. Was 0.3llllllllllqsn8td1p464unmi.
FAIL Number.prototype.toString.call(new Number(0.1), 36) should be 0.3lllllllllm. Was 0.3llllllllllqsn8td1p464unmi.
PASS Number(-1.1).toString() is "-1.1"
PASS Number.prototype.toString.call(-1.1) is "-1.1"
PASS Number.prototype.toString.call(new Number(-1.1)) is "-1.1"
PASS Number("-1.1").toString() is "-1.1"
PASS Number(-1.1).toString(10) is "-1.1"
PASS Number(-1.1).toString(2) is "-1.000110011001100110011001100110011001100110011001101"
PASS Number.prototype.toString.call(-1.1, 2) is "-1.000110011001100110011001100110011001100110011001101"
PASS Number.prototype.toString.call(new Number(-1.1), 2) is "-1.000110011001100110011001100110011001100110011001101"
FAIL Number(-1.1).toString(36) should be -1.3llllllllm. Was -1.3lllllllllxagau2ctidswz5mi.
FAIL Number.prototype.toString.call(-1.1, 36) should be -1.3llllllllm. Was -1.3lllllllllxagau2ctidswz5mi.
FAIL Number.prototype.toString.call(new Number(-1.1), 36) should be -1.3llllllllm. Was -1.3lllllllllxagau2ctidswz5mi.
PASS Number(1.1).toString() is "1.1"
PASS Number.prototype.toString.call(1.1) is "1.1"
PASS Number.prototype.toString.call(new Number(1.1)) is "1.1"
PASS Number("1.1").toString() is "1.1"
PASS Number(1.1).toString(10) is "1.1"
PASS Number(1.1).toString(2) is "1.000110011001100110011001100110011001100110011001101"
PASS Number.prototype.toString.call(1.1, 2) is "1.000110011001100110011001100110011001100110011001101"
PASS Number.prototype.toString.call(new Number(1.1), 2) is "1.000110011001100110011001100110011001100110011001101"
FAIL Number(1.1).toString(36) should be 1.3llllllllm. Was 1.3lllllllllxagau2ctidswz5mi.
FAIL Number.prototype.toString.call(1.1, 36) should be 1.3llllllllm. Was 1.3lllllllllxagau2ctidswz5mi.
FAIL Number.prototype.toString.call(new Number(1.1), 36) should be 1.3llllllllm. Was 1.3lllllllllxagau2ctidswz5mi.
PASS Number(1984.1).toString() is "1984.1"
PASS Number.prototype.toString.call(1984.1) is "1984.1"
PASS Number.prototype.toString.call(new Number(1984.1)) is "1984.1"
PASS Number("1984.1").toString() is "1984.1"
PASS Number(1984.1).toString(10) is "1984.1"
PASS Number(1984.1).toString(2) is "11111000000.00011001100110011001100110011001100110011"
PASS Number.prototype.toString.call(1984.1, 2) is "11111000000.00011001100110011001100110011001100110011"
PASS Number.prototype.toString.call(new Number(1984.1), 2) is "11111000000.00011001100110011001100110011001100110011"
FAIL Number(1984.1).toString(36) should be 1j4.3lllllllc. Was 1j4.3lllllllcd2obsszcl3di.
FAIL Number.prototype.toString.call(1984.1, 36) should be 1j4.3lllllllc. Was 1j4.3lllllllcd2obsszcl3di.
FAIL Number.prototype.toString.call(new Number(1984.1), 36) should be 1j4.3lllllllc. Was 1j4.3lllllllcd2obsszcl3di.
PASS Number(-1984.1).toString() is "-1984.1"
PASS Number.prototype.toString.call(-1984.1) is "-1984.1"
PASS Number.prototype.toString.call(new Number(-1984.1)) is "-1984.1"
PASS Number("-1984.1").toString() is "-1984.1"
PASS Number(-1984.1).toString(10) is "-1984.1"
PASS Number(-1984.1).toString(2) is "-11111000000.00011001100110011001100110011001100110011"
PASS Number.prototype.toString.call(-1984.1, 2) is "-11111000000.00011001100110011001100110011001100110011"
PASS Number.prototype.toString.call(new Number(-1984.1), 2) is "-11111000000.00011001100110011001100110011001100110011"
FAIL Number(-1984.1).toString(36) should be -1j4.3lllllllc. Was -1j4.3lllllllcd2obsszcl3di.
FAIL Number.prototype.toString.call(-1984.1, 36) should be -1j4.3lllllllc. Was -1j4.3lllllllcd2obsszcl3di.
FAIL Number.prototype.toString.call(new Number(-1984.1), 36) should be -1j4.3lllllllc. Was -1j4.3lllllllcd2obsszcl3di.
PASS Number(2147483647.1).toString() is "2147483647.1"
PASS Number.prototype.toString.call(2147483647.1) is "2147483647.1"
PASS Number.prototype.toString.call(new Number(2147483647.1)) is "2147483647.1"
PASS Number("2147483647.1").toString() is "2147483647.1"
PASS Number(2147483647.1).toString(10) is "2147483647.1"
PASS Number(2147483647.1).toString(2) is "1111111111111111111111111111111.000110011001100110011"
PASS Number.prototype.toString.call(2147483647.1, 2) is "1111111111111111111111111111111.000110011001100110011"
PASS Number.prototype.toString.call(new Number(2147483647.1), 2) is "1111111111111111111111111111111.000110011001100110011"
FAIL Number(2147483647.1).toString(36) should be zik0zj.3lllg. Was zik0zj.3lllfu07ldi.
FAIL Number.prototype.toString.call(2147483647.1, 36) should be zik0zj.3lllg. Was zik0zj.3lllfu07ldi.
FAIL Number.prototype.toString.call(new Number(2147483647.1), 36) should be zik0zj.3lllg. Was zik0zj.3lllfu07ldi.
PASS Number(-2147483648.1).toString() is "-2147483648.1"
PASS Number.prototype.toString.call(-2147483648.1) is "-2147483648.1"
PASS Number.prototype.toString.call(new Number(-2147483648.1)) is "-2147483648.1"
PASS Number("-2147483648.1").toString() is "-2147483648.1"
PASS Number(-2147483648.1).toString(10) is "-2147483648.1"
PASS Number(-2147483648.1).toString(2) is "-10000000000000000000000000000000.000110011001100110011"
PASS Number.prototype.toString.call(-2147483648.1, 2) is "-10000000000000000000000000000000.000110011001100110011"
PASS Number.prototype.toString.call(new Number(-2147483648.1), 2) is "-10000000000000000000000000000000.000110011001100110011"
FAIL Number(-2147483648.1).toString(36) should be -zik0zk.3lllg. Was -zik0zk.3lllfu07ldi.
FAIL Number.prototype.toString.call(-2147483648.1, 36) should be -zik0zk.3lllg. Was -zik0zk.3lllfu07ldi.
FAIL Number.prototype.toString.call(new Number(-2147483648.1), 36) should be -zik0zk.3lllg. Was -zik0zk.3lllfu07ldi.
PASS Number(9007199254740992).toString() is "9007199254740992"
PASS Number.prototype.toString.call(9007199254740992) is "9007199254740992"
PASS Number.prototype.toString.call(new Number(9007199254740992)) is "9007199254740992"
PASS Number("9007199254740992").toString() is "9007199254740992"
PASS Number(9007199254740992).toString(10) is "9007199254740992"
PASS Number(9007199254740992).toString(2) is "100000000000000000000000000000000000000000000000000000"
PASS Number.prototype.toString.call(9007199254740992, 2) is "100000000000000000000000000000000000000000000000000000"
PASS Number.prototype.toString.call(new Number(9007199254740992), 2) is "100000000000000000000000000000000000000000000000000000"
PASS Number(9007199254740992).toString(36) is "2gosa7pa2gw"
PASS Number.prototype.toString.call(9007199254740992, 36) is "2gosa7pa2gw"
PASS Number.prototype.toString.call(new Number(9007199254740992), 36) is "2gosa7pa2gw"
PASS Number(-9007199254740992).toString() is "-9007199254740992"
PASS Number.prototype.toString.call(-9007199254740992) is "-9007199254740992"
PASS Number.prototype.toString.call(new Number(-9007199254740992)) is "-9007199254740992"
PASS Number("-9007199254740992").toString() is "-9007199254740992"
PASS Number(-9007199254740992).toString(10) is "-9007199254740992"
PASS Number(-9007199254740992).toString(2) is "-100000000000000000000000000000000000000000000000000000"
PASS Number.prototype.toString.call(-9007199254740992, 2) is "-100000000000000000000000000000000000000000000000000000"
PASS Number.prototype.toString.call(new Number(-9007199254740992), 2) is "-100000000000000000000000000000000000000000000000000000"
PASS Number(-9007199254740992).toString(36) is "-2gosa7pa2gw"
PASS Number.prototype.toString.call(-9007199254740992, 36) is "-2gosa7pa2gw"
PASS Number.prototype.toString.call(new Number(-9007199254740992), 36) is "-2gosa7pa2gw"
PASS successfullyParsed is true

TEST COMPLETE
